+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/dto/product/ProductResponse.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 @Data
 public class ProductResponse {
-    private Long id;
+    private Short id;
     private String name;
     private String description;
     private int price;
     private String status;
     private String imageURL;
-    private Long categoryId;
+    private Short categoryId;
     private List<ProductRating> ratings;
 }
