+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/dto/product/ProductSearchRequest.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 @Data
 public class ProductSearchRequest {
-    private List<Long> categories;
+    private List<Short> categories;
 }
