+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/repository/SliderRepository.java
@@ -3,5 +3,5 @@ package by.dz.fruits_in_choco.fruits_in_choco.repository;
 import by.dz.fruits_in_choco.fruits_in_choco.entity.slide.Slide;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SliderRepository extends JpaRepository<Slide, Long> {
+public interface SliderRepository extends JpaRepository<Slide, Short> {
 }
