+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/OrderService.java
@@ -6,9 +6,9 @@ import java.util.List;
 import java.util.Map;
 
 public interface OrderService {
-    Order makeOrder(Order order, Map<Long, Integer> products);
+    Order makeOrder(Order order, Map<Short, Short> products);
     List<Order> getAllOrders();
-    Order getOrderById(Long id);
-    void deleteOrderById(Long id);
-    Order updateOrder(Order order, Long id);
+    Order getOrderById(Short id);
+    void deleteOrderById(Short id);
+    Order updateOrder(Order order, Short id);
 }
