+++ src/main/java/by/dz/fruits_in_choco/fruits_in_choco/service/UserService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface UserService {
     List<User> getUsers(int page, int size, String direction, String sortBy);
-    User getUserById(Long id);
+    User getUserById(Short id);
     User getUserByEmail(String email);
 }
